+++ src/globals/workingManikins.ts
@@ -18,6 +18,7 @@ import { nginxManikin } from '../mobparts/manikins/nginx';
 import { piholeManikin } from '../mobparts/manikins/pihole';
 import { portainerManikin } from '../mobparts/manikins/portainer';
 import { rocketchatManikin } from '../mobparts/manikins/rocketchat';
+import { styleguidistManikin } from '../mobparts/manikins/styleguidist';
 import { swarmpitManikin } from '../mobparts/manikins/swarmpit';
 import { traefikManikin } from '../mobparts/manikins/traefik';
 import { visualizerManikin } from '../mobparts/manikins/visualizer';
@@ -56,6 +57,7 @@ export const workingManikins: IManikin[] = [
   piholeManikin,
   portainerManikin,
   rocketchatManikin,
+  styleguidistManikin,
   swarmpitManikin,
   traefikManikin,
   tvManikin,