+++ src/mobparts/mites/network/styleguidistNetworkMite.ts
@@ -1,0 +1,16 @@
+import { IMite } from '../../../interfaces/IMite';
+
+export const styleguidistNetworkMite: IMite = {
+  type: `DockerSwarmNetwork`,
+  miteIndex: 1022,
+  miteString: `
+
+# Begin Styleguidist Network Section
+
+ styleguidist:
+  driver: overlay
+
+# End Styleguidist Network Section
+
+`,
+};