+++ src/mobparts/mites/network/tvNetworkMite.ts
@@ -8,7 +8,7 @@ import { IMite } from '../../../interfaces/IMite';
 
 export const tvNetworkMite: IMite = {
   type: `DockerSwarmNetwork`,
-  miteIndex: 1024,
+  miteIndex: 1025,
   miteString: `
 
 # Begin TV Network Section