+++ src/mobparts/mites/network/webdavNetworkMite.ts
@@ -2,7 +2,7 @@ import { IMite } from '../../../interfaces/IMite';
 
 export const webdavNetworkMite: IMite = {
   type: `DockerSwarmNetwork`,
-  miteIndex: 1027,
+  miteIndex: 1028,
   miteString: `
 
 #Begin Webdav Network Section