+++ src/mobparts/mites/service/tvServiceMite.ts
@@ -8,7 +8,7 @@ import { IMite } from '../../../interfaces/IMite';
 
 export const tvServiceMite: IMite = {
   type: `DockerSwarmService`,
-  miteIndex: 2025,
+  miteIndex: 2026,
   miteString: `
 
 # Begin TV Service Section